requirements.txt
openai>=1.3.0
pyperclip>=1.8.2
keyboard>=0.13.5
pillow>=9.0.0
